:truck: Distribute `functions` and `setup` methods into specific files
