+++ lib/class-template-loader.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Handles loading template files or whatever: Template_Loader class.
+ *
+ * @package HrswpTheme
+ * @since 2.0.0
+ */
+
+namespace HrswpTheme\lib\Template_Loader;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	die( 'Silence is golden.' );
+}
+
+/**
+ * The HRSWP Theme Template Loader class.
+ */
+class Template_Loader {
+	// Coming soon.
+}
