+++ lib/global-styles.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * Future home of global styles management using json imports maybe
+ *
+ * @package HrswpTheme
+ * @since 2.0.0
+ */
+
+// Colors import/export.
+
+// Font import/export.
